Issue #36
Fixed I2C_MODES Datatypes from 16 to 32 bit

+++ src/MCP7940.cpp
@@ -228,7 +228,7 @@ MCP7940_Class::~MCP7940_Class() {} // of class destructor                     //
 ** Method begin starts I2C communications with the device, using a default address if one is not specified and    **
 ** return true if the device has been detected and false if it was not                                            **
 *******************************************************************************************************************/
-bool MCP7940_Class::begin(const uint16_t i2cSpeed) {                          // Start I2C communications         //
+bool MCP7940_Class::begin(const uint32_t i2cSpeed) {                          // Start I2C communications         //
   Wire.begin();                                                               // Start I2C as master device       //
   Wire.setClock(i2cSpeed);                                                    // Set the I2C bus speed            //
   Wire.beginTransmission(MCP7940_ADDRESS);                                    // Address the MCP7940M             //
